--- cupertino/src/commonMain/kotlin/io/github/alexzhirkevich/cupertino/swipebox/DragAnchors.kt
@@ -1,11 +1,0 @@
-package io.github.alexzhirkevich.cupertino.swipebox
-
-enum class DragAnchors {
-    Start,
-    Center,
-    End,
-    StartToEndDismissed,
-    StartToEndExpanded,
-    EndToStartDismissed,
-    EndToStartExpanded
-}
