--- cupertino/src/commonMain/kotlin/io/github/alexzhirkevich/cupertino/swipebox/SwipeDirection.kt
@@ -1,5 +1,0 @@
-package io.github.alexzhirkevich.cupertino.swipebox
-
-enum class SwipeDirection {
-    StartToEnd, EndToStart, Both
-}
